Updated with some perf numbers based on the latest iteration.

@Tornhoof - I updated the `JumpTableMultipleEntryBenchmark` to use different lengths of strings 👍 


## TLDR
- The IL Emit trie is the best data structure we have matching strings with N >=2
- The overall result is about a 35% speedup in our real world benchmarks
- The vectorized version (64bit) has enough scalability to dominate the conversation for all reasonable sizes
- The vectorized version is faster for small numbers of strings
- The non-vectorized version (32bit) is better for most sizes but doesn't scale as well (N >=~50)
- The IL Emit trie is slower than our current implementation for a single string (N == 1)
- I have a prototype vectorized non-IL-Emit implementation that's much better than our current implementation, but will not be following up on it in this PR. It's possible that once I productionize it we'll end up in the same place. 



## JumpTableMultipleEntryBenchmark

|       Method | Count |       Mean |     Error |     StdDev |     Median |          Op/s | Scaled | ScaledSD | Allocated |
|------------- |------ |-----------:|----------:|-----------:|-----------:|--------------:|-------:|---------:|----------:|
|     **Baseline** |     **2** |   **9.157 ns** | **0.0119 ns** |  **0.0093 ns** |   **9.154 ns** | **109,206,844.3** |   **1.00** |     **0.00** |       **0 B** |
| LinearSearch |     2 |  24.462 ns | 0.4854 ns |  1.1536 ns |  23.687 ns |  40,879,272.5 |   2.67 |     0.13 |       0 B |
|   Dictionary |     2 | 239.305 ns | 4.7769 ns |  9.9711 ns | 233.227 ns |   4,178,763.6 |  26.13 |     1.08 |       0 B |
|         Trie |     2 |  12.490 ns | 0.1453 ns |  0.1135 ns |  12.442 ns |  80,064,498.0 |   1.36 |     0.01 |       0 B |
|   VectorTrie |     2 |  12.239 ns | 0.4438 ns |  0.3934 ns |  12.061 ns |  81,709,153.3 |   1.34 |     0.04 |       0 B |
|              |       |            |           |            |            |               |        |          |           |
|     **Baseline** |     **5** |   **9.451 ns** | **0.1893 ns** |  **0.4075 ns** |   **9.215 ns** | **105,810,490.7** |   **1.00** |     **0.00** |       **0 B** |
| LinearSearch |     5 |  59.557 ns | 0.3104 ns |  0.2424 ns |  59.485 ns |  16,790,717.5 |   6.31 |     0.26 |       0 B |
|   Dictionary |     5 | 250.408 ns | 5.4164 ns | 15.7998 ns | 249.498 ns |   3,993,477.5 |  26.54 |     1.99 |       0 B |
|         Trie |     5 |  15.329 ns | 0.0352 ns |  0.0294 ns |  15.326 ns |  65,234,792.0 |   1.62 |     0.07 |       0 B |
|   VectorTrie |     5 |  14.448 ns | 0.2894 ns |  0.6934 ns |  13.977 ns |  69,215,327.3 |   1.53 |     0.10 |       0 B |
|              |       |            |           |            |            |               |        |          |           |
|     **Baseline** |    **10** |   **9.184 ns** | **0.0564 ns** |  **0.0471 ns** |   **9.170 ns** | **108,890,713.8** |   **1.00** |     **0.00** |       **0 B** |
| LinearSearch |    10 | 109.005 ns | 0.4852 ns |  0.3788 ns | 108.837 ns |   9,173,927.8 |  11.87 |     0.07 |       0 B |
|   Dictionary |    10 | 233.432 ns | 2.2215 ns |  1.7344 ns | 232.913 ns |   4,283,900.9 |  25.42 |     0.22 |       0 B |
|         Trie |    10 |  20.698 ns | 0.0195 ns |  0.0163 ns |  20.697 ns |  48,312,758.8 |   2.25 |     0.01 |       0 B |
|   VectorTrie |    10 |  18.475 ns | 0.3657 ns |  0.8621 ns |  18.035 ns |  54,127,825.9 |   2.01 |     0.09 |       0 B |
|              |       |            |           |            |            |               |        |          |           |
|     **Baseline** |    **25** |   **9.179 ns** | **0.0447 ns** |  **0.0349 ns** |   **9.173 ns** | **108,940,288.1** |   **1.00** |     **0.00** |       **0 B** |
| LinearSearch |    25 | 253.955 ns | 1.8727 ns |  1.6601 ns | 253.249 ns |   3,937,709.1 |  27.67 |     0.20 |       0 B |
|   Dictionary |    25 | 251.812 ns | 4.9460 ns |  8.1265 ns | 248.261 ns |   3,971,222.8 |  27.43 |     0.88 |       0 B |
|         Trie |    25 | 130.515 ns | 3.1445 ns |  4.8020 ns | 128.177 ns |   7,661,961.4 |  14.22 |     0.52 |       0 B |
|   VectorTrie |    25 |  28.468 ns | 0.7217 ns |  0.7411 ns |  28.150 ns |  35,127,220.9 |   3.10 |     0.08 |       0 B |
|              |       |            |           |            |            |               |        |          |           |
|     **Baseline** |    **50** |   **9.162 ns** | **0.0149 ns** |  **0.0132 ns** |   **9.158 ns** | **109,152,276.2** |   **1.00** |     **0.00** |       **0 B** |
| LinearSearch |    50 | 422.275 ns | 8.2855 ns | 14.9405 ns | 412.976 ns |   2,368,125.0 |  46.09 |     1.61 |       0 B |
|   Dictionary |    50 | 239.422 ns | 3.3859 ns |  3.0015 ns | 238.496 ns |   4,176,717.7 |  26.13 |     0.32 |       0 B |
|         Trie |    50 | 253.350 ns | 2.1008 ns |  1.8623 ns | 252.924 ns |   3,947,102.3 |  27.65 |     0.20 |       0 B |
|   VectorTrie |    50 |  45.970 ns | 0.9156 ns |  2.2460 ns |  44.898 ns |  21,753,348.8 |   5.02 |     0.24 |       0 B |
|              |       |            |           |            |            |               |        |          |           |
|     **Baseline** |   **100** |   **9.398 ns** | **0.1872 ns** |  **0.3694 ns** |   **9.215 ns** | **106,404,380.9** |   **1.00** |     **0.00** |       **0 B** |
| LinearSearch |   100 | 574.650 ns | 0.7291 ns |  0.6089 ns | 574.604 ns |   1,740,188.6 |  61.23 |     2.25 |       0 B |
|   Dictionary |   100 | 240.242 ns | 0.9625 ns |  0.7515 ns | 239.967 ns |   4,162,464.6 |  25.60 |     0.94 |       0 B |
|         Trie |   100 | 498.034 ns | 1.9317 ns |  1.5081 ns | 497.815 ns |   2,007,893.2 |  53.07 |     1.96 |       0 B |
|   VectorTrie |   100 | 159.908 ns | 3.1971 ns |  8.2528 ns | 154.723 ns |   6,253,581.5 |  17.04 |     1.08 |       0 B |

## JumpTableSingleEntryBenchmark

|         Method |      Mean |     Error |    StdDev |    Median |          Op/s | Scaled | ScaledSD | Allocated |
|--------------- |----------:|----------:|----------:|----------:|--------------:|-------:|---------:|----------:|
|       Baseline |  8.986 ns | 0.0478 ns | 0.0373 ns |  8.976 ns | 111,281,592.4 |   1.00 |     0.00 |       0 B |
| Implementation |  8.322 ns | 0.1743 ns | 0.2865 ns |  8.179 ns | 120,158,894.2 |   0.93 |     0.03 |       0 B |
|      Prototype |  5.752 ns | 0.2388 ns | 0.2117 ns |  5.689 ns | 173,865,156.4 |   0.64 |     0.02 |       0 B |
|           Trie | 12.045 ns | 0.2485 ns | 0.5131 ns | 11.718 ns |  83,022,932.1 |   1.34 |     0.06 |       0 B |
|     VectorTrie | 11.729 ns | 0.3663 ns | 0.3762 ns | 11.621 ns |  85,261,040.1 |   1.31 |     0.04 |       0 B |


## Azure

**Before**

|   Method |       Mean |    Error |   StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|--------- |-----------:|---------:|---------:|------------:|-------:|---------:|-------:|----------:|
| Baseline |   643.8 ns | 20.28 ns | 19.92 ns | 1,553,224.9 |   1.00 |     0.00 |      - |   3.91 KB |
|      Dfa | 2,203.0 ns | 41.75 ns | 46.40 ns |   453,932.3 |   3.42 |     0.12 | 0.7324 |   62.7 KB |

**After**

|   Method |       Mean |    Error |   StdDev |     Median |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|--------- |-----------:|---------:|---------:|-----------:|------------:|-------:|---------:|-------:|----------:|
| Baseline |   656.0 ns | 12.95 ns | 26.46 ns |   644.3 ns | 1,524,287.5 |   1.00 |     0.00 |      - |   3.91 KB |
|      Dfa | 1,846.8 ns | 35.37 ns | 37.85 ns | 1,839.7 ns |   541,487.6 |   2.82 |     0.12 | 0.4883 |  45.13 KB |


## Github

**Before**

|   Method |     Mean |     Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|--------- |---------:|----------:|----------:|------------:|-------:|---------:|-------:|----------:|
| Baseline | 252.4 ns |  1.932 ns |  1.713 ns | 3,962,251.3 |   1.00 |     0.00 | 0.1221 |   9.49 KB |
|      Dfa | 926.7 ns | 21.847 ns | 21.457 ns | 1,079,059.1 |   3.67 |     0.09 | 0.9766 |  87.58 KB |

**After**

|   Method |     Mean |      Error |     StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|--------- |---------:|-----------:|-----------:|------------:|-------:|---------:|-------:|----------:|
| Baseline | 248.7 ns |  0.8757 ns |  0.7763 ns | 4,020,625.9 |   1.00 |     0.00 |      - |   9.49 KB |
|      Dfa | 719.5 ns | 14.2196 ns | 15.8050 ns | 1,389,924.7 |   2.89 |     0.06 | 0.7324 |  73.51 KB |