@rynowak 2018-08-17
Putting this out for discussion... obviously not finished, but seems functionally correct.

## AND FAST

|           Method | Count |       Mean |     Error |    StdDev |          Op/s | Scaled | ScaledSD | Allocated |
|----------------- |------ |-----------:|----------:|----------:|--------------:|-------:|---------:|----------:|
|         Baseline |    25 |   5.444 ns | 0.0739 ns | 0.0655 ns | 183,674,816.5 |   1.00 |     0.00 |       0 B |
|     LinearSearch |    25 | 290.574 ns | 5.6966 ns | 6.5602 ns |   3,441,462.7 |  53.38 |     1.32 |       0 B |
|       Dictionary |    25 | 165.146 ns | 3.1412 ns | 3.2258 ns |   6,055,261.4 |  30.34 |     0.67 |       0 B |
|            Ascii |    25 |         NA |        NA |        NA |            NA |      ? |        ? |       N/A |
| DictionaryLookup |    25 |  26.721 ns | 0.4535 ns | 0.4020 ns |  37,424,116.5 |   4.91 |     0.09 |       0 B |
|  CustomHashTable |    25 |  20.253 ns | 0.3149 ns | 0.2945 ns |  49,376,385.9 |   3.72 |     0.07 |       0 B |
|             Trie |    25 |  15.233 ns | 0.3414 ns | 0.5610 ns |  65,646,955.5 |   2.80 |     0.11 |       0 B |