@JamesNK 2018-08-17
I think it is cool and worth exploring. Blows away my idea of just grouping by the length!

I worry about how much IL will be generated though.