@rynowak 2018-08-17
Yeah, keep in mind that this is at a 'show and tell' level of quality right now 😆 

Numbers for this are very very very good:

## Github

**Before**

|   Method |     Mean |     Error |   StdDev |   Median |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|--------- |---------:|----------:|---------:|---------:|------------:|-------:|---------:|-------:|----------:|
| Baseline | 153.1 ns |  3.006 ns | 5.719 ns | 150.2 ns | 6,530,888.5 |   1.00 |     0.00 | 0.0610 |   9.49 KB |
|      Dfa | 571.9 ns | 10.056 ns | 8.398 ns | 571.3 ns | 1,748,555.9 |   3.74 |     0.14 | 0.9766 |  87.58 KB |

**After**

|   Method |     Mean |    Error |    StdDev |        Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|--------- |---------:|---------:|----------:|------------:|-------:|---------:|-------:|----------:|
| Baseline | 158.5 ns | 3.005 ns |  2.811 ns | 6,310,600.4 |   1.00 |     0.00 | 0.1221 |   9.49 KB |
|      Dfa | 393.2 ns | 7.817 ns | 10.164 ns | 2,543,193.3 |   2.48 |     0.08 | 0.8545 |  73.51 KB |