@rynowak 2018-08-17
re: IL 

I'm also getting stoked on the idea of of vectorizing. This would look like reading, normalizing and comparing 4 characters at a time. This will really help with the amount of codegen, because you're doing a lot more with less.