@damageboy 2018-08-17
@rynowak sorry if I'm jumping into the middle of the discussion, so this is basically "switching" on the length, then on str[0], str[1] etc. right? I'm a bit slow in reading IL, just thought I'd ask before asking something else..