@Tornhoof 2018-08-17
Nice approach @rynowak.
Anyway, I tried porting my json member name matching code from SpanJson to Jumptable: 
https://gist.github.com/Tornhoof/e88bf7ea470d315372229c427efb2652 (case-sensitive). It tries to do a more optimized comparison as described here:
https://github.com/Tornhoof/SpanJson/wiki/Technology-and-Internals#3-comparison-against-integers

But it looks like your approach is faster anyway, my approach is slower (up to twice as slow according to the benchmark) , probably due to the high overhead of calling the delegate for each match (the code was designed to be called in a large loop for deserialization).
Maybe this gives you further ideas for improvements.