@benaadams 2018-08-17
`uint` and `ulong` ascii casing https://github.com/GrabYourPitchforks/coreclr/blob/a3c1df25c4225995ffd6b18fd0fc39d6b81fd6a5/src/System.Private.CoreLib/src/System/Text/Utf16Utility.cs courtesy of @GrabYourPitchforks