@rynowak 2018-08-17
@damageboy 

> so this is basically "switching" on the length, then on str[0], str[1] etc. right? I'm a bit slow in reading IL, just thought I'd ask before asking something else

yes, thats exactly what this does. 


@Tornhoof - that's pretty rad. @davidfowl  and I were discussing vectorizing this and it looks like you've already done that. 

> Probably due to the high overhead of calling the delegate for each match

Curious which delegate is this that's being called? It seems to me like you might have already built what I was thinking about.  

Looking more closely I think the difference between your approach and what I was imagining is that you're using a flat list of cases, where I was planning to build a 'tree' - resulting in each region of memory being read and considered in sequence. I need this to scale to a high number of entries that you wouldn't find when looking at class members. 

@benaadams cool beans 👍 https://github.com/GrabYourPitchforks/coreclr/blob/a3c1df25c4225995ffd6b18fd0fc39d6b81fd6a5/src/System.Private.CoreLib/src/System/Text/Utf16Utility.cs#L81 is exactly like what I was planning. 


---------------

I assume that @davidfowl must have tweeted this lol. I'm glad to see all of the interest. 

What I have in mind for the next iteration of this is to turn it into something like the following (psuedocode)

```
uint64 bits;
var span = path.AsSpan(segment.Start, segment.Length);
switch (span.Length)
{
    case 8:
        bits = ReadUInt64(span, 0);
        bits = ToLowerInvariantAsciiQWord(bits);
        
        if (bits == 0xSomePrecomputed64bitNumber)
        {
            bits = ReadUInt64(span, 4);
            bits = ToLowerInvariantAsciiQWord(bits);

            if (bits == 0xTheNextPrecomputed64bitNumber)
            {
                return 5; // Return the state that matches this entry
            }

            return _defaultDesintation;
        }

        if (bits == 0xSomeOtherPrecomputed64bitNumber)
        {
            ...
        }

        ....
        return _defaultDesintation;

    case 17:
     .....
}
```