@rynowak 2018-08-17
Also for anyone who's not familiar with how we're testing this - we've built a tool that can convert swagger files to benchmarks for routing. 
https://github.com/aspnet/Routing/tree/release/2.2/benchmarks/Swaggatherer
https://github.com/aspnet/Routing/blob/release/2.2/benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherGithubBenchmark.generated.cs (warning BIG)
https://github.com/aspnet/Routing/blob/release/2.2/benchmarks/Microsoft.AspNetCore.Routing.Performance/Matching/MatcherAzureBenchmark.generated.cs (warning BIGGER)